% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_dtw_clustering.R
\name{seq_dtw_clustering}
\alias{seq_dtw_clustering}
\title{Create clusters using DTW}
\usage{
seq_dtw_clustering(
  eventlog,
  type,
  k,
  ...,
  preproc,
  distance,
  centroid,
  control,
  args,
  seed,
  trace,
  error.check
)
}
\arguments{
\item{eventlog}{An event log object containing case, activity and timestamp
classifiers.}

\item{type}{What type of clustering method to use: \code{"partitional"}, \code{"hierarchical"}, \code{"tadpole"}
or \code{"fuzzy"}.}

\item{k}{Number of desired clusters. It can be a numeric vector with different values.}

\item{...}{Arguments to pass to preprocessing, centroid \strong{and} distance functions (added to
\code{args}). Also passed to \code{method} from \code{\link[dtwclust:tsclust-controls]{hierarchical_control()}} if it happens to be a function,
and to \code{\link[stats:hclust]{stats::hclust()}} if it contains the \code{members} parameter.}

\item{preproc}{Function to preprocess data. Defaults to \code{\link[dtwclust:zscore]{zscore()}} \emph{only} if \code{centroid} \code{=}
\code{"shape"}, but will be replaced by a custom function if provided.}

\item{distance}{A registered distance from \code{\link[proxy:dist]{proxy::dist()}}. Ignored for \code{type} \code{=} \code{"tadpole"}.}

\item{centroid}{Either a supported string, or an appropriate function to calculate centroids when
using partitional/hierarchical/tadpole methods. See Centroids section.}

\item{control}{An appropriate list of controls. See \link[dtwclust]{tsclust-controls}.}

\item{args}{An appropriate list of arguments for preprocessing, distance and centroid functions.
See \code{\link[dtwclust:tsclust-controls]{tsclust_args()}} and the examples.}

\item{seed}{Random seed for reproducibility.}

\item{trace}{Logical flag. If \code{TRUE}, more output regarding the progress is printed to screen.}

\item{error.check}{Logical indicating whether the function should try to detect inconsistencies
and give more informative errors messages. Also used internally to avoid repeating checks.}
}
\value{
A list containing the clusters assigned to each case and an object
  with an appropriate class from \link[dtwclust]{TSClusters}.
}
\description{
Create clusters using Dynamic Time Warping (DTW) by providing
  an event log object.
}
\examples{
\dontrun{
library(seqClustR)
eventlog <- bupaR::activities_to_eventlog(data_frame,
                       case_id,
                       activity_id,
                       timestamp)
seq_dtw_clustering(eventlog)
}
}
\seealso{
\link[bupaR]{eventlog}, \link[dtwclust]{tsclust}
}
