% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_edit_distance_clustering.R
\name{seq_edit_distance_clustering}
\alias{seq_edit_distance_clustering}
\title{Create clusters using Edit Distance}
\usage{
seq_edit_distance_clustering(eventlog, method, weight, q, p, bt, nthread)
}
\arguments{
\item{eventlog}{An event log object containing case, activity and timestamp
classifiers.}

\item{method}{Distance metric. See \code{\link[stringdist]{stringdist-metrics}}}

\item{weight}{For \code{method='osa'} or \code{'dl'}, the penalty for
deletion, insertion, substitution and transposition, in that order. When
\code{method='lv'}, the penalty for transposition is ignored. When
\code{method='jw'}, the weights associated with characters of \code{a},
characters from \code{b} and the transposition weight, in that order. 
Weights must be positive and not exceed 1. \code{weight} is ignored
completely when \code{method='hamming'}, \code{'qgram'}, \code{'cosine'},
\code{'Jaccard'}, or \code{'lcs'}}

\item{q}{Size of the \eqn{q}-gram; must be nonnegative. Only applies to
\code{method='qgram'}, \code{'jaccard'} or \code{'cosine'}.}

\item{p}{Prefix factor for Jaro-Winkler distance. The valid range for 
\code{p} is \code{0 <= p <= 0.25}. If \code{p=0} (default), the
Jaro-distance is returned. Applies only to \code{method='jw'}.}

\item{bt}{Winkler's boost threshold. Winkler's prefix factor is
only applied when the Jaro distance is larger than \code{bt}
Applies only to \code{method='jw'} and \code{p>0}.}

\item{nthread}{Maximum number of threads to use. By default, a sensible
number of threads is chosen, see \code{\link[stringdist]{stringdist-parallelization}}.}
}
\value{
A list containing the clusters assigned to each case and the fitted
  Hierarchical Clustering model object.
}
\description{
Create clusters using Edit Distance metric and Hierarchical
  Clustering by providing an event log object.
}
\examples{
\dontrun{
library(seqClustR)
eventlog <- bupaR::activities_to_eventlog(data_frame,
                       case_id,
                       activity_id,
                       timestamp)
seq_edit_distance_clustering(eventlog)
}
}
\seealso{
\link[bupaR]{eventlog}, \link[stringdist]{seq_distmatrix}, \link[stats]{hclust}
}
