% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_kmeans_clustering.R
\name{seq_kmeans_clustering}
\alias{seq_kmeans_clustering}
\title{Create clusters using K-Means Method}
\usage{
seq_kmeans_clustering(
  eventlog,
  centers = 2,
  iter.max = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)
}
\arguments{
\item{eventlog}{An event log object containing case, activity and timestamp
classifiers.}

\item{centers}{either the number of clusters, say \eqn{k}, or a set of
    initial (distinct) cluster centres.  If a number, a random set of
    (distinct) rows in \code{x} is chosen as the initial centres.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if \code{centers} is a number, how many random sets
    should be chosen?}

\item{algorithm}{character: may be abbreviated.  Note that
    \code{"Lloyd"} and \code{"Forgy"} are alternative names for one
    algorithm.}

\item{trace}{logical or integer number, currently only used in the
    default method (\code{"Hartigan-Wong"}): if positive (or true),
    tracing information on the progress of the algorithm is
    produced.  Higher values may produce more tracing information.}
}
\value{
A list containing the clusters assigned to each case and the fitted
  K-Means model.
}
\description{
Create clusters using K-Means Method by providing an event log
  object.
}
\examples{
\dontrun{
library(seqClustR)
eventlog <- bupaR::activities_to_eventlog(data_frame,
                       case_id,
                       activity_id,
                       timestamp)
seq_kmeans_clustering(eventlog)
}
}
\seealso{
\link[bupaR]{eventlog}, \link[stats]{kmeans}
}
