% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_markov_clustering.R
\name{seq_markov_clustering}
\alias{seq_markov_clustering}
\title{Create clusters using Markov Model}
\usage{
seq_markov_clustering(
  eventlog,
  informat,
  alphabet,
  states,
  id,
  weights,
  start,
  left,
  right,
  gaps,
  missing,
  void,
  nr,
  cnames,
  xtstep,
  tick.last,
  cpal,
  missing.color,
  labels,
  n_clusters,
  transition_probs,
  initial_probs,
  formula,
  data,
  coefficients,
  cluster_names,
  em_step,
  global_step,
  local_step,
  control_em,
  control_global,
  control_local,
  lb,
  ub,
  threads,
  log_space,
  ...
)
}
\arguments{
\item{eventlog}{An event log object containing case, activity and timestamp
classifiers.}

\item{n_clusters}{A scalar giving the number of clusters/submodels 
(not used if starting values for model parameters are given with 
\code{initial_probs} and \code{transition_probs}).}

\item{transition_probs}{A list of matrices of transition
probabilities for submodels of each cluster.}

\item{initial_probs}{A list which contains vectors of initial state
probabilities for submodels of each cluster.}

\item{formula}{Covariates as an object of class \code{\link{formula}},
left side omitted.}

\item{data}{An optional data frame, list or environment containing the variables
in the model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{coefficients}{An optional \eqn{k x l} matrix of regression coefficients for 
time-constant covariates for mixture probabilities, where \eqn{l} is the number 
of clusters and \eqn{k} is the number of covariates. A logit-link is used for
mixture probabilities. The first column is set to zero.}

\item{cluster_names}{A vector of optional names for the clusters.}

\item{em_step}{Logical. Whether or not to use the EM algorithm at the start
of the parameter estimation. The default is \code{TRUE}.}

\item{global_step}{Logical. Whether or not to use global optimization via
\code{\link{nloptr}} (possibly after the EM step). The default is \code{FALSE}.}

\item{local_step}{Logical. Whether or not to use local optimization via
\code{\link{nloptr}} (possibly after the EM and/or global steps). The default is \code{FALSE}.}

\item{control_em}{Optional list of control parameters for the EM algorithm.
Possible arguments are \describe{
\item{maxeval}{The maximum number of iterations, the default is 1000. 
Note that iteration counter starts with -1 so with \code{maxeval=1} you get already two iterations.
This is for backward compatibility reasons.}
\item{print_level}{The level of printing. Possible values are 0
(prints nothing), 1 (prints information at the start and the end of the algorithm),
2 (prints at every iteration),
and for mixture models 3 (print also during optimization of coefficients).}
\item{reltol}{Relative tolerance for convergence defined as
\eqn{(logLik_new - logLik_old)/(abs(logLik_old) + 0.1)}.
The default is 1e-10.}
\item{restart}{A list containing options for possible EM restarts with the
  following components:
\describe{
\item{times}{Number of restarts of the EM algorithm using random initial values. The default is 0, i.e. no restarts. }
\item{transition}{Logical. Should the original transition probabilities be varied? The default is \code{TRUE}. }
\item{emission}{Logical. Should the original emission probabilities be varied? The default is \code{TRUE}. }
\item{sd}{Standard deviation for \code{rnorm} used in randomization. The default is 0.25.}
\item{maxeval}{Maximum number of iterations, the default is \code{control_em$maxeval}}
\item{print_level}{Level of printing in restarted EM steps. The default is \code{control_em$print_level}. }
\item{reltol}{Relative tolerance for convergence at restarted EM steps. The default is \code{control_em$reltol}.
If the relative change of the final model of the restart phase is larger than the tolerance
for the original EM phase, the final model is re-estimated with the original \code{reltol}
and \code{maxeval} at the end of the EM step.}
\item{n_optimum}{Save the log-likelihood values of the \code{n_optimum} best
models (from all estimated models including the the first EM run.).
The default is \code{min(times + 1, 25)}.}
\item{use_original}{If \code{TRUE}. Use the initial values of the input model as starting 
points for the permutations. Otherwise permute the results of the first EM run.}
}
}
}}

\item{control_global}{Optional list of additional arguments for
  \code{\link{nloptr}} argument \code{opts}. The default values are
  \describe{
   \item{algorithm}{\code{"NLOPT_GD_MLSL_LDS"}}
   \item{local_opts}{\code{list(algorithm = "NLOPT_LD_LBFGS", ftol_rel = 1e-6, xtol_rel = 1e-4)}}
   \item{maxeval}{\code{10000} (maximum number of iterations in global optimization algorithm.)}
   \item{maxtime}{\code{60} (maximum time for global optimization. Set to 0 for unlimited time.)}
}}

\item{control_local}{Optional list of additional arguments for
\code{\link{nloptr}} argument \code{opts}. The default values are
\describe{
 \item{algorithm}{\code{"NLOPT_LD_LBFGS"}}
 \item{ftol_rel}{\code{1e-10}}
 \item{xtol_rel}{\code{1e-8}}
 \item{maxeval}{\code{10000} (maximum number of iterations)}
}}

\item{lb}{Lower and upper bounds for parameters in Softmax parameterization.
The default interval is \eqn{[pmin(-25, 2*initialvalues), pmax(25, 2*initialvalues)]},
except for gamma coefficients,
where the scale of covariates is taken into account.
Note that it might still be a good idea to scale covariates around unit scale.
Bounds are used only in the global optimization step.}

\item{ub}{Lower and upper bounds for parameters in Softmax parameterization.
The default interval is \eqn{[pmin(-25, 2*initialvalues), pmax(25, 2*initialvalues)]},
except for gamma coefficients,
where the scale of covariates is taken into account.
Note that it might still be a good idea to scale covariates around unit scale.
Bounds are used only in the global optimization step.}

\item{threads}{Number of threads to use in parallel computing. The default is 1.}

\item{log_space}{Make computations using log-space instead of scaling for greater
numerical stability at a cost of decreased computational performance. The default is \code{FALSE}.}

\item{...}{Additional arguments to \code{simulate_transition_probs}.}
}
\value{
A list containing the clusters assigned to each case and the fitted
  Markov Model.
}
\description{
Create clusters using Markov Model by providing an event log
  object.
}
\examples{
\dontrun{
library(seqClustR)
eventlog <- bupaR::activities_to_eventlog(data_frame,
                       case_id,
                       activity_id,
                       timestamp)
seq_markov_clustering(eventlog)
}
}
\seealso{
\link[bupaR]{eventlog}, \link[seqHMM]{seqdef}, \link[seqHMM]{build_mmm}, \link[seqHMM]{fit_model}
}
