% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_event_log.R
\name{split_event_log}
\alias{split_event_log}
\title{Get event logs for each cluster}
\usage{
split_event_log(eventlog, cluster_assignment)
}
\arguments{
\item{eventlog}{An event log object containing case, activity and timestamp
classifiers.}

\item{cluster_assignment}{A data frame containing the clusters assigned
for each case. Output of seq_edit_distance_clustering,
seq_kmeans_clustering, seq_markov_clustering or seq_dtw_clustering can be
passed here.}
}
\value{
A list containing event logs for each cluster.
}
\description{
This function helps in getting event logs for each cluster which
  can then be used to create process models for clusters and visualize them
  as well using the fuzzymineR package or by using some other tools like
  ProM or Disco.
}
\examples{
\dontrun{
library(seqClustR)
eventlog <- bupaR::activities_to_eventlog(data_frame,
                       case_id,
                       activity_id,
                       timestamp)
split_event_log(eventlog, cluster_assignment)
}
}
\seealso{
\link[bupaR]{eventlog}, \link[bupaR]{activities_to_eventlog}
}
